feat: index ahora lang es español. Agrego mas sugerencias
