+++ index.html
@@ -1,9 +1,10 @@
 <!doctype html>
-<html lang="en">
+<html lang="es">
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta name="description" content="TODO: agregar descripcion util" />
     <title>Facundo Aylan</title>
   </head>
   <body>
